int answers converted to double ➿

+++ src/main/java/school/mjc/stage0/conditions/task3/SalaryCalculator.java
@@ -5,11 +5,11 @@ public class SalaryCalculator {
         if (salary < 0) {
             System.out.println("wrong input!");
         } else if (salary < 10000) {
-            System.out.println(salary - salary * 15 / 100);
+            System.out.println((double) (salary - salary * 15 / 100));
         } else if (salary > 10000 && salary <= 20000) {
-            System.out.println(salary - salary * 18 / 100);
+            System.out.println((double) (salary - salary * 18 / 100));
         } else if (salary > 20000) {
-            System.out.println(salary - salary * 20 / 100);
+            System.out.println((double) (salary - salary * 20 / 100));
         }
     }
 }
